refactor(capture-sdk): Delete example-app-shared-code
Moved the example-app-shared-code files to screen-api-example-app. The package was not needed anymore because we had a single example app (before there was another example app in the capture-sdk).
