--- capture-sdk/example-app-shared-code/gradle.properties
@@ -1,1 +1,0 @@
-android.useAndroidX=true
