+++ capture-sdk/screen-api-example-app/src/main/java/net/gini/android/capture/screen/ExampleUtil.java
@@ -1,4 +1,4 @@
-package net.gini.android.capture.example.shared;
+package net.gini.android.capture.screen;
 
 import android.content.Intent;
 import android.os.Bundle;
