+++ capture-sdk/screen-api-example-app/src/main/java/net/gini/android/capture/screen/ExtractionsActivity.kt
@@ -10,7 +10,6 @@ import android.view.View
 import android.view.ViewGroup
 import android.view.inputmethod.InputMethodManager
 import android.widget.Toast
-import androidx.activity.OnBackPressedCallback
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.FileProvider
 import androidx.core.widget.addTextChangedListener
@@ -21,7 +20,6 @@ import com.google.android.material.textfield.TextInputLayout
 import net.gini.android.capture.Amount
 import net.gini.android.capture.AmountCurrency
 import net.gini.android.capture.GiniCapture
-import net.gini.android.capture.example.shared.BaseExampleApp
 import net.gini.android.capture.network.GiniCaptureDefaultNetworkService
 import net.gini.android.capture.network.model.GiniCaptureCompoundExtraction
 import net.gini.android.capture.network.model.GiniCaptureSpecificExtraction
