+++ capture-sdk/screen-api-example-app/src/main/java/net/gini/android/capture/screen/RuntimePermissionHandler.java
@@ -1,4 +1,4 @@
-package net.gini.android.capture.example.shared;
+package net.gini.android.capture.screen;
 
 import android.Manifest;
 import android.annotation.SuppressLint;
